+++ Source/Scene/CameraFlightPath.js
@@ -197,9 +197,9 @@ define([
     var scratchStartRight = new Cartesian3();
     var currentFrame = new Matrix4();
 
-    function createUpdate3D(frameState, destination, duration, direction, up) {
-        var camera = frameState.camera;
-        var ellipsoid = frameState.mapProjection.ellipsoid;
+    function createUpdate3D(scene, destination, duration, direction, up) {
+        var camera = scene.camera;
+        var ellipsoid = scene.mapProjection.ellipsoid;
 
         var start = camera.cameraToWorldCoordinatesPoint(camera.position, scratchStartPosition);
         var startDirection = camera.cameraToWorldCoordinatesVector(camera.direction, scratchStartDirection);
@@ -318,9 +318,9 @@ define([
         });
     }
 
-    function createUpdateCV(frameState, destination, duration, direction, up) {
-        var camera = frameState.camera;
-        var ellipsoid = frameState.mapProjection.ellipsoid;
+    function createUpdateCV(scene, destination, duration, direction, up) {
+        var camera = scene.camera;
+        var ellipsoid = scene.mapProjection.ellipsoid;
 
         var path = createPath2D(camera, ellipsoid, Cartesian3.clone(camera.position), destination, duration);
         var orientations = createOrientations2D(camera, path, direction, up);
@@ -344,9 +344,9 @@ define([
         return update;
     }
 
-    function createUpdate2D(frameState, destination, duration, direction, up) {
-        var camera = frameState.camera;
-        var ellipsoid = frameState.mapProjection.ellipsoid;
+    function createUpdate2D(scene, destination, duration, direction, up) {
+        var camera = scene.camera;
+        var ellipsoid = scene.mapProjection.ellipsoid;
 
         var start = Cartesian3.clone(camera.position);
         start.z = camera.frustum.right - camera.frustum.left;
@@ -422,7 +422,7 @@ define([
         }
         //>>includeEnd('debug');
 
-        if (scene.frameState.mode === SceneMode.MORPHING) {
+        if (scene.mode === SceneMode.MORPHING) {
             return {
                 duration : 0
             };
@@ -439,9 +439,8 @@ define([
 
         var convert = defaultValue(options.convert, true);
 
-        var frameState = scene.frameState;
-        if (convert && frameState.mode !== SceneMode.SCENE3D) {
-            var projection = frameState.mapProjection;
+        if (convert && scene.mode !== SceneMode.SCENE3D) {
+            var projection = scene.mapProjection;
             var ellipsoid = projection.ellipsoid;
             ellipsoid.cartesianToCartographic(destination, scratchCartographic);
             destination = projection.project(scratchCartographic, scratchDestination);
@@ -464,21 +463,22 @@ define([
         var onComplete = wrapCallback(options.onComplete);
         var onCancel = wrapCallback(options.onCancel);
 
+        var camera = scene.camera;
         var referenceFrame = options.endReferenceFrame;
         if (defined(referenceFrame)) {
-            scene.camera.setTransform(referenceFrame);
+            camera.setTransform(referenceFrame);
         }
 
-        var frustum = frameState.camera.frustum;
-        if (frameState.mode === SceneMode.SCENE2D) {
-            if (Cartesian2.equalsEpsilon(frameState.camera.position, destination, CesiumMath.EPSILON6) && (CesiumMath.equalsEpsilon(Math.max(frustum.right - frustum.left, frustum.top - frustum.bottom), destination.z, CesiumMath.EPSILON6))) {
+        var frustum = camera.frustum;
+        if (scene.mode === SceneMode.SCENE2D) {
+            if (Cartesian2.equalsEpsilon(camera.position, destination, CesiumMath.EPSILON6) && (CesiumMath.equalsEpsilon(Math.max(frustum.right - frustum.left, frustum.top - frustum.bottom), destination.z, CesiumMath.EPSILON6))) {
                 return {
                     duration : 0,
                     onComplete : onComplete,
                     onCancel: onCancel
                 };
             }
-        } else if (Cartesian3.equalsEpsilon(destination, frameState.camera.position, CesiumMath.EPSILON6)) {
+        } else if (Cartesian3.equalsEpsilon(destination, camera.position, CesiumMath.EPSILON6)) {
             return {
                 duration : 0,
                 onComplete : onComplete,
@@ -489,7 +489,7 @@ define([
         if (duration <= 0) {
             var newOnComplete = function() {
                 var position = destination;
-                if (frameState.mode === SceneMode.SCENE3D) {
+                if (scene.mode === SceneMode.SCENE3D) {
                     if (!defined(options.direction) && !defined(options.up)){
                         dirScratch = Cartesian3.normalize(Cartesian3.negate(position, dirScratch), dirScratch);
                         rightScratch = Cartesian3.normalize(Cartesian3.cross(dirScratch, Cartesian3.UNIT_Z, rightScratch), rightScratch);
@@ -509,13 +509,13 @@ define([
                     upScratch = defaultValue(options.up, Cartesian3.cross(rightScratch, dirScratch, upScratch));
                 }
 
-                Cartesian3.clone(position, frameState.camera.position);
-                Cartesian3.clone(dirScratch, frameState.camera.direction);
-                Cartesian3.clone(upScratch, frameState.camera.up);
-                Cartesian3.clone(rightScratch, frameState.camera.right);
+                Cartesian3.clone(position, camera.position);
+                Cartesian3.clone(dirScratch, camera.direction);
+                Cartesian3.clone(upScratch, camera.up);
+                Cartesian3.clone(rightScratch, camera.right);
 
-                if (frameState.mode === SceneMode.SCENE2D) {
-                    var zoom = frameState.camera.position.z;
+                if (scene.mode === SceneMode.SCENE2D) {
+                    var zoom = camera.position.z;
                     var ratio = frustum.top / frustum.right;
 
                     var incrementAmount = (zoom - (frustum.right - frustum.left)) * 0.5;
@@ -537,12 +537,12 @@ define([
         }
 
         var update;
-        if (frameState.mode === SceneMode.SCENE3D) {
-            update = createUpdate3D(frameState, destination, duration, direction, up);
-        } else if (frameState.mode === SceneMode.SCENE2D) {
-            update = createUpdate2D(frameState, destination, duration, direction, up);
+        if (scene.mode === SceneMode.SCENE3D) {
+            update = createUpdate3D(scene, destination, duration, direction, up);
+        } else if (scene.mode === SceneMode.SCENE2D) {
+            update = createUpdate2D(scene, destination, duration, direction, up);
         } else {
-            update = createUpdateCV(frameState, destination, duration, direction, up);
+            update = createUpdateCV(scene, destination, duration, direction, up);
         }
 
         return {
@@ -583,18 +583,14 @@ define([
         if (!defined(scene)) {
             throw new DeveloperError('scene is required.');
         }
-        if (!defined(scene.frameState)) {
-            throw new DeveloperError('frameState is required.');
-        }
+
         if (!defined(rectangle)) {
             throw new DeveloperError('options.destination is required.');
         }
         //>>includeEnd('debug');
 
-        var frameState = scene.frameState;
         var createAnimationoptions = clone(options);
-        var camera = frameState.camera;
-        camera.getRectangleCameraCoordinates(rectangle, c3destination);
+        scene.camera.getRectangleCameraCoordinates(rectangle, c3destination);
 
         createAnimationoptions.destination = c3destination;
         createAnimationoptions.convert = false;